requirements.txt
# Aircraft Engine Monitoring System - Python Dependencies

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization and Dashboards
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0

# Machine Learning (for future development)
scikit-learn>=1.3.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Streaming and Message Queue
kafka-python>=2.0.2

# Database Connectivity
psycopg2-binary>=2.9.7
redis>=4.6.0

# Data Analysis and Statistics
scipy>=1.11.0
statsmodels>=0.14.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development and Testing (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
